New RNG engine for MaNGOS
  - Based on C++11 <random> library
  - Uses ACE_TSS_Singleton for thread-safety

+++ src/game/Server/WorldSocket.cpp
@@ -89,7 +89,7 @@ WorldSocket::WorldSocket(void) :
     m_OutBufferLock(),
     m_OutBuffer(0),
     m_OutBufferSize(65536),
-    m_Seed(static_cast<uint32>(rand32()))
+    m_Seed(rand32())
 {
     reference_counting_policy().value(ACE_Event_Handler::Reference_Counting_Policy::ENABLED);
 }
